+++ src/features/Feedback/index.ts
@@ -1,3 +1,4 @@
-export { default as MentorQuarterlyFeedbackQuestionnaire } from './MentorQuarterlyFeedbackQuestionnaire'
+export { default as MentorQuarterlyFeedbackQuestionnaire } from './questionnaires/MentorQuarterlyFeedbackQuestionnaire'
+export { default as MenteeQuarterlyFeedbackQuestionnaire } from './questionnaires/MenteeQuarterlyFeedbackQuestionnaire'
 export { default as MenteeFeedback } from './MenteeFeedback'
 export { default as MentorFeedback } from './MentorFeedback'
